Apply dhewm3 / b9435a432db5d0d6d4a78141dc56d27b75bdd792: Fix CRC32 hashing for x86_64
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6206 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ idlib/Dict.cpp
@@ -249,7 +249,7 @@ idDict::Checksum
 ================
 */
 int	idDict::Checksum( void ) const {
-	unsigned long ret;
+	unsigned int ret;
 
 	idList<idKeyValue> sorted = args;
 	sorted.Sort( KeyCompare );
